rename, delete redundant file
